Add command line arguments
This commit refactors the main code and allows passing in a few command
line arguments to modify the functionality without having to hard code
them in the script. Some of this functionality might be, perhaps, better
expressed as a separate object, but for now this seems to work just
fine.