Split sbatch template into two
This commit splits out the payload from the sbatch template. Since that portion
is common to most simulations, I thought it was better to just remove it from
the main template so that each simulation can have separate sbatch settings but
run the same payload (since most of the payload options can be determined
directly from the case setup).