Add python file to read and write OF solutions
This python module provides functions to read an OpenFOAM solution into
memory for further processing, as well as to write data back out as an
OpenFOAM solution. My hope is to use this to be able to do some
preprocessing of the data and write it back out to OpenFOAM format to be
able to visualize it with paraFoam.