Compute and save the number of cells in the domain
I was having some trouble with preparing OpenFOAM simulation data for
clustering. The issue was that some fields can be uniform valued fields
which then need to be expanded into a vector with the same
dimensionality as the other fields (given by the number of cells).
However, since this information is not known up front, it was causing
issues in how I had to figure it out. So I figured I should just compute
it when reading the OpenFOAM simulation initially and saving it with the
pickled data (as a metadata field).

This did require changes in how the pickled data is stored (and
therefore how it is processed further). This commit introduces all of
these changes.